Add mtdev-git lorry
On request by Joe Burmeister.

This is meant to replace the existing mtdev repository, which imports
a tarball. The old repository does not seem to be in use, but I don't
know if it may be in use by some older release, a personal branch, or
some other morphs repository, so I do not wish to meddle with th eold
repository. The old repo should have been called mtdev-tarball, but
wasn't, so I'm calling the new one mtdev-git instead.