Add U-Boot repository to support Hardkernel Odroid C1
This adds a lorry for the Hardkernel U-Boot tree for the Odroid C1
development board. This is required since this board uses a new core
(Cortex-A5) and SoC (Amlogic Meson S802) and there is a massive amount
of platform specific support added by the OEM in this non-upstreamed
tree. Currently it would be impractical to port the vast number of
changes to a modern U-Boot version. This repo also contains a
binary-only bootloader which is required to boot the board.

Change-Id: I8f0f7b1930e65d725a4de34d9cfbd631c0f70500